+++ CHANGELOG.rst
@@ -1,4 +1,10 @@
-v1.10.0 (2023-12-23)
+v1.11.0 (2023-12-25)
+*******************
+* Acquisition and AutoFocus both got a `broadcast` option to disable broadcast of images.
+* AutoFocus got a `final_image` parameter to take a final image at optimal focus.
+
+
+v1.10.0 (2023-12-24)
 *******************
 * Added CallModule script.
 * Changed ScriptRunner module so that it can run a script multiple times.